Add files via upload
Sometimes we just want a quick fast tool to really tell whats the current update, we just need a bare minimum of data. Web scraping deals with taking some data from the web and then processing it and displaying the relevant content in a short and crisp manner. 
What the code is doing ? 
 

First we are using Tkinter Library to make GUI required for our script
We are using requests Library to get the data from the unofficial api
Then we are displaying the data we need in this case its Total active cases: and confirmed cases